lint fix: renamed Icon -> CustomIcon; workflow working dir fix
