+++ frontend/prototype/src/components/CustomIcon.vue
